+++ OUDS/Core/Themes/Orange/Sources/OrangeTheme.swift
@@ -12,7 +12,7 @@
 //
 
 import Foundation
-import OUDSThemesCommons
+import OUDS
 
 /// This is an override of the default basic `OUDSTheme` and should be seen as the default theme for the OUDS library.
 /// It can overrides any properties from its superclass, and can be derived too.